# next.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inbound API v2",
    "version": "2.0.0",
    "description": "The official API documentation for Inbound - Email API Service. Manage domains, send emails, handle webhooks, and more.",
    "contact": {
      "name": "Inbound Support",
      "url": "https://inbound.new/support",
      "email": "support@inbound.new"
    },
    "license": {
      "name": "Commercial",
      "url": "https://inbound.new/terms"
    }
  },
  "servers": [
    {
      "url": "https://inbound.new/api/v2",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api/v2",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API Key authentication. Use 'Bearer YOUR_API_KEY'"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for session-based authentication"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Authentication required. Please provide a valid API key or session token."
                },
                "code": {
                  "type": "string",
                  "example": "401"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred. Please try again later."
                },
                "code": {
                  "type": "string",
                  "example": "500"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/v2/endpoints": {
      "get": {
        "operationId": "get-v2-endpoints",
        "summary": "Get all endpoints",
        "description": "Retrieve all endpoints for the authenticated user with filtering, sorting, and pagination options.",
        "tags": [
          "Endpoints"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "number"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "number"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "webhook",
                "email",
                "email_group"
              ]
            },
            "required": false
          },
          {
            "in": "query",
            "name": "active",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "oldest"
              ]
            },
            "required": false
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of endpoints with statistics and pagination",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "webhook",
                              "email",
                              "email_group"
                            ]
                          },
                          "config": {},
                          "isActive": {
                            "type": "boolean"
                          },
                          "description": {
                            "type": "string",
                            "nullable": true
                          },
                          "userId": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "groupEmails": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            },
                            "nullable": true
                          },
                          "deliveryStats": {
                            "type": "object",
                            "properties": {
                              "total": {
                                "type": "number"
                              },
                              "successful": {
                                "type": "number"
                              },
                              "failed": {
                                "type": "number"
                              },
                              "lastDelivery": {
                                "type": "string",
                                "nullable": true
                              }
                            }
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "limit": {
                          "type": "number"
                        },
                        "offset": {
                          "type": "number"
                        },
                        "total": {
                          "type": "number"
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-v2-endpoints",
        "summary": "Create a new endpoint",
        "description": "Create a new webhook, email forward, or email group endpoint for processing incoming emails.",
        "tags": [
          "Endpoints"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "nullable": false
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "webhook",
                      "email",
                      "email_group"
                    ],
                    "nullable": false
                  },
                  "config": {
                    "nullable": false
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Endpoint created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "webhook",
                        "email",
                        "email_group"
                      ]
                    },
                    "config": {},
                    "isActive": {
                      "type": "boolean"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "userId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "groupEmails": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "nullable": true
                    },
                    "deliveryStats": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "successful": {
                          "type": "number"
                        },
                        "failed": {
                          "type": "number"
                        },
                        "lastDelivery": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}